Move hook_entity_prepare_form() implementation from farm_log to farm_log_asset.
This hook pertains to populating the asset reference field on
logs, which is provided by the farm_log_asset module.
